button colo

+++ client/src/pages/batch/displayallbatchdetails.js
@@ -2,7 +2,6 @@ import React, { Component } from 'react';
 import API from '../../API/Batch';
 import BatchRecord from './Getbatchdetails';
 import BatchInfo from './BatchInfo';
-import Homepage from "../general/Homepage";
 import { connect } from 'react-redux';
 import moment from "moment";
 
@@ -56,53 +55,63 @@ class Allbatches extends Component {
 
 
   loadBatches = () => {
-    if (this.props.usertype === "management") {
-      API.getAllBatch()
-        .then(response => {
-          //console.log("The Batch Details of  - axios call", response.data);
-          let batchrecords = [];
-          if (response.data.length > 0) {
-            for (let i = 0; i < response.data.length; i++) {
-              // console.log("Records", response.data[i]._id, response.data[i].batchdesc, response.data[i].level, response.data[i].teacher);
-              let currentrec = {
-                recid: response.data[i]._id,
-                recdesc: response.data[i].batchdesc,
-                recsubj: response.data[i].course,
-                reclevel: response.data[i].level,
-                teacher: response.data[i].teacherName,
-                teacher_id: response.data[i]._id,
-                noofstu: response.data[i].students.length,
-                noofclasses: response.data[i].classid.length,
-                examDate: moment(response.data[i].examDate).format("MM-DD-YYYY")
-                //recstudents: response.data[i].students
-              }
-              batchrecords.push(currentrec);
-            } // end for
+    // if (this.props.usertype !== "management") {
+    //   console.log("Console error", this.props.usertype, "----------------------------")
+    //   return; //(<Homepage msg="Please Login" />);
+    // }
 
-            this.setState({
-              batchrecords: batchrecords,
-              allbatches: true
-            })
-            console.log("Batch list", batchrecords)
-            // ,() => { console.log("State of records") });
-          } else {
-            this.setState({
-              batchrecords: [],
-              allbatches: false
-            });
-            console.log("Axios reponse -No Cohorts exist");
-          }
-        }) // end then
-        .catch(error => {
-          this.setState({ allbatches: false })
-          console.log("Error in getting batch records!!!", error);
-          // return (<Homepage msg="Please Login" />);
-        });
-    }
-    else {
-      console.log("Console error", this.props.usertype, "----------------------------")
-      return (<Homepage msg="Please Login" />);
-    }
+    API.getAllBatch()
+      .then(response => {
+        //console.log("The Batch Details of  - axios call", response.data);
+        let batchrecords = [];
+        if (response.data.length > 0) {
+          for (let i = 0; i < response.data.length; i++) {
+            const rec = response.data[i]
+            const teacherDoc = rec.teacher;
+            const teacherName =
+              teacherDoc && typeof teacherDoc === "object"
+                ? `${teacherDoc.fname || ""} ${teacherDoc.lname || ""}`.trim()
+                : rec.teacherName || "Not Assigned";
+            // console.log("Records", response.data[i]._id, response.data[i].batchdesc, response.data[i].level, response.data[i].teacher);
+            let currentrec = {
+              recid: rec._id,
+              recdesc: rec.batchdesc,
+              recsubj: rec.course,
+              reclevel: rec.level,
+              teacher: teacherName,
+              teacher_id:
+                teacherDoc && typeof teacherDoc === "object" && teacherDoc._id
+                  ? String(teacherDoc._id)
+                  : "", 
+              noofstu: rec.students?.length,
+              noofclasses: rec.classid?.length,
+              examDate: rec.examDate ? moment(rec.examDate).format("MM-DD-YYYY") : ""
+              //recstudents: response.data[i].students
+            }
+            batchrecords.push(currentrec);
+          } // end for
+
+          this.setState({
+            batchrecords: batchrecords,
+            allbatches: true
+          })
+          console.log("Batch list", batchrecords)
+          // ,() => { console.log("State of records") });
+        } else {
+          this.setState({
+            batchrecords: [],
+            allbatches: false
+          });
+          console.log("Axios reponse -No Cohorts exist");
+        }
+      }) // end then
+      .catch(error => {
+        this.setState({ allbatches: false })
+        console.log("Error in getting batch records!!!", error);
+        // return (<Homepage msg="Please Login" />);
+      });
+
+
   }
 
   showAllBatches = () => {
@@ -127,6 +136,7 @@ class Allbatches extends Component {
   }
 
   deleteBatch = (batchid) => {
+    if(!batchid)return;
     console.log("delete", batchid)
     API.deleteBatch(batchid)
       .then(response => {
@@ -145,8 +155,8 @@ class Allbatches extends Component {
             level: '',
             course: '',
             examDate: ''
-         }
-       });
+          }
+        });
       })
   }
 
@@ -154,7 +164,7 @@ class Allbatches extends Component {
   render() {
     const stbatchrec = this.state.batchrecords;
     // console.log("Display all batch details --",this.props);
-    if (this.props.usertype === "management") {
+    //if (this.props.usertype === "management") {
       return (<div className="container">
         <div className="middlecontent">
           {this.state.batchSelected ? (
@@ -202,11 +212,11 @@ class Allbatches extends Component {
       </div> //End container
       ); // end return
     }// end if props.usertype is boardmember
-    else {
-      return (<Homepage msg="Please Login as Board member to see all batches" />);
-    }
+    //else {
+    //  return (<Homepage msg="Please Login as Board member to see all batches" />);
+    //}
   } //end render
-} //end allbatches
+//} //end allbatches
 
 
 const mapStateToProps = (state) => {
